requirements.txt
beautifulsoup4==4.6.3
lxml==4.2.5
attrs==18.2.0
